test: adding test for RegisterInterchainAccount & adding check to relay_test
