+++ modules/apps/27-interchain-accounts/host/keeper/account_test.go
@@ -1,0 +1,33 @@
+package keeper_test
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	"github.com/cosmos/ibc-go/v2/modules/apps/27-interchain-accounts/types"
+	ibctesting "github.com/cosmos/ibc-go/v2/testing"
+)
+
+func (suite *KeeperTestSuite) TestRegisterInterchainAccount() {
+	suite.SetupTest()
+
+	path := NewICAPath(suite.chainA, suite.chainB)
+	suite.coordinator.SetupConnections(path)
+
+	// InitInterchainAccount
+	err := SetupICAPath(path, TestOwnerAddress)
+	suite.Require().NoError(err)
+
+	portId, err := types.GeneratePortID(TestOwnerAddress, ibctesting.FirstConnectionID, ibctesting.FirstConnectionID)
+	suite.Require().NoError(err)
+
+	// Get the address of the interchain account stored in state during handshake step
+	storedAddr, found := suite.chainB.GetSimApp().ICAHostKeeper.GetInterchainAccountAddress(suite.chainB.GetContext(), portId)
+	suite.Require().True(found)
+
+	icaAddr, err := sdk.AccAddressFromBech32(storedAddr)
+	suite.Require().NoError(err)
+
+	// Check if account is created
+	interchainAccount := suite.chainB.GetSimApp().AccountKeeper.GetAccount(suite.chainB.GetContext(), icaAddr)
+	suite.Require().Equal(interchainAccount.GetAddress().String(), storedAddr)
+}
