+++ modules/apps/27-interchain-accounts/host/keeper/handshake_test.go
@@ -168,7 +168,6 @@ func (suite *KeeperTestSuite) TestOnChanOpenTry() {
 			} else {
 				suite.Require().Error(err)
 			}
-
 		})
 	}
 }
